Removed more things, general cleanup.  Added new script file, but it isn't being used yet.

+++ Sources/Model/NBody/Simulation/Core/Data/Random/NBodySimulationRandom.mm
@@ -96,7 +96,7 @@ void Data::Random::acquire(GLfloat* pPosition,
     // set our global points and velocity pointers, etc.
     gPoints = pPosition;
     gVelocities = pVelocity;
-    gParticleCount = mnBodies;
+    gParticleCount = static_cast<unsigned int>(mnBodies);
     
     std::string fullpath;
     CF::IFStreamRef pStream = CF::IFStreamCreate(CFSTR("bang"), CFSTR("lua"), &fullpath);
